fix(ci): split results_to_df module and fix ruff format
Move DataFrame export helpers to results_df.py to keep utils.py under
pylint's 1000-line limit. Share OVERALL_LABEL via constants.py to avoid
cyclic imports. Re-export results_to_df from utils via __getattr__.
