+++ seametrics/tracking/constants.py
@@ -1,0 +1,3 @@
+"""Shared constants for tracking metrics."""
+
+OVERALL_LABEL = "OVERALL"
