+++ seametrics/tracking/utils.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 import contextlib
 import io
 import pathlib
-from typing import Protocol
 
 import numpy as np
 import pandas as pd
@@ -13,10 +12,7 @@ from tqdm import tqdm
 
 from ._box_utils import box_convert, box_denormalize
 
-#: Label used for the pooled, dataset-level aggregate row appended by
-#: :func:`results_to_df`. Matches the name ``motmetrics`` uses for its overall
-#: row so MOT and HOTA DataFrames stay consistent.
-OVERALL_LABEL = "OVERALL"
+#: Re-exported from :mod:`constants` for backward compatibility.
 
 #: Metric names aggregated by summation across sequences (integer counts), as
 #: opposed to ratio/derived metrics which are pooled. Shared so the metric
@@ -36,18 +32,6 @@ COUNT_METRICS = (
 #: Column width for MOT tracker-format arrays produced by
 #: :func:`prepare_data_for_det_metrics`.
 _TRACKER_ARRAY_COLS = 10
-
-
-class _MetricsForDf(Protocol):
-    """Structural type for :func:`results_to_df` metric instances."""
-
-    accumulators: dict[str, object]
-    comparison_excluded: frozenset[str]
-    RESULT_LAYOUT: str
-
-    def compute(self, sequence: list[str] | str | None = ...) -> dict: ...
-
-    def compute_many(self, sequence_list: list[str]) -> dict: ...
 
 
 def _as_tracker_array(rows: list) -> np.ndarray:
@@ -814,20 +798,6 @@ def compute_all_metrics_by_sequence(
     return instances
 
 
-def _sequence_list_for_df(
-    metrics: _MetricsForDf,
-    sequence_list: list[str] | None,
-) -> list[str]:
-    """Resolve per-sequence rows to include before pooling OVERALL."""
-    if sequence_list is not None:
-        return sequence_list
-    names = list(metrics.accumulators.keys())
-    excluded = metrics.comparison_excluded
-    if not excluded:
-        return names
-    return [name for name in names if name not in excluded]
-
-
 def compute_sizes(view: fo.DatasetView, gt_field: str) -> list:
     """Compute bounding-box areas for all annotated objects in a sequence view.
 
@@ -917,113 +887,6 @@ def get_sequence_info(
     return sequence_info
 
 
-def _scale_metric_row(flat_result: dict, *, layout: str) -> dict:
-    """Apply metric-specific scaling to a flat ``{metric: scalar}`` result.
-
-    TrackingMetrics: ``mota`` is scaled x100 and ``motp`` is converted to
-    ``(1 - motp) x 100``; all other metrics are left unchanged.
-    HOTAMetrics: all metric values (hota, deta, assa, loca) are scaled x100,
-    except the ``num_unique_objects`` count.
-
-    Args:
-        flat_result: Mapping from metric name to a single scalar value.
-        layout: ``"flat"`` for HOTA-style results, ``"nested"`` for MOT.
-
-    Returns:
-        New dict with scaling applied.
-    """
-    if layout == "flat":
-        return {
-            k: (v if k == "num_unique_objects" else v * 100)
-            for k, v in flat_result.items()
-        }
-    row = dict(flat_result)
-    row["mota"] *= 100
-    row["motp"] = (1 - row["motp"]) * 100
-    return row
-
-
-def _flatten_result(result: dict, key: str) -> dict:
-    """Flatten a ``compute()`` result to a flat ``{metric: scalar}`` mapping.
-
-    Nested motmetrics-style results (TrackingMetrics and HOTA ``compute_many``)
-    map *key* to each metric's inner entry — ``OVERALL_LABEL`` for the pooled
-    row, or a sequence name for a per-sequence row.
-
-    Args:
-        result: Raw dict returned by ``metrics.compute(...)`` or
-            ``metrics.compute_many(...)``.
-        key: Inner key to select for nested results (sequence name or
-            ``OVERALL_LABEL``).
-
-    Returns:
-        Flat ``{metric: scalar}`` dict.
-    """
-    return {k: v[key] for k, v in result.items()}
-
-
-def _compute_table_bundle(metrics: _MetricsForDf, sequence_list: list) -> dict:
-    """Return one nested result dict covering every table row."""
-    if metrics.RESULT_LAYOUT == "flat":
-        return metrics.compute_many(sequence_list)
-    return metrics.compute(sequence=sequence_list)
-
-
-def results_to_df(
-    metrics: _MetricsForDf, sequence_list: list | None = None
-) -> pd.DataFrame:
-    """Convert TrackingMetrics or HOTAMetrics results to a DataFrame.
-
-    One batched ``compute()`` / ``compute_many()`` call returns all per-sequence
-    rows and the pooled OVERALL row. Appends a pooled OVERALL row — the
-    MOT-standard dataset-level score, not a mean of per-sequence values.
-
-    Args:
-        metrics: Fitted metric instance with ``accumulators`` and ``compute()``.
-        sequence_list: Sequence names to include. Defaults to all accumulators
-            minus :attr:`~TrackingMetrics.comparison_excluded` (set by
-            :func:`compute_all_metrics_by_sequence` for fair cross-model pooling).
-
-    Raises:
-        ValueError: If *sequence_list* contains the reserved ``OVERALL`` label.
-    """
-    sequence_list = _sequence_list_for_df(metrics, sequence_list)
-
-    if not sequence_list:
-        return pd.DataFrame()
-
-    if OVERALL_LABEL in sequence_list:
-        raise ValueError(f"{OVERALL_LABEL!r} is reserved for pooled results")
-
-    layout = metrics.RESULT_LAYOUT
-    bundle = _compute_table_bundle(metrics, sequence_list)
-    rows = []
-    for sequence in sequence_list:
-        row = _scale_metric_row(_flatten_result(bundle, key=sequence), layout=layout)
-        row["sequence"] = sequence
-        rows.append(row)
-    row = _scale_metric_row(_flatten_result(bundle, key=OVERALL_LABEL), layout=layout)
-    row["sequence"] = OVERALL_LABEL
-    rows.append(row)
-
-    return pd.DataFrame(rows)
-
-
-def hota_results_to_df(
-    metrics: _MetricsForDf, sequence_list: list | None = None
-) -> pd.DataFrame:
-    """Alias for results_to_df for backward compatibility.
-
-    Args:
-        metrics: Fitted metric instance (see :func:`results_to_df`).
-        sequence_list: Optional list of sequence names to include.
-
-    Returns:
-        DataFrame with one row per sequence and one column per metric value.
-    """
-    return results_to_df(metrics, sequence_list)
-
-
 def classify_num_objects(x: int | float) -> str | None:
     """Map an object count to a human-readable category label.
 
@@ -1047,3 +910,19 @@ def classify_num_objects(x: int | float) -> str | None:
             category = label
             break
     return category
+
+
+def __getattr__(name: str):
+    if name == "OVERALL_LABEL":
+        from .constants import OVERALL_LABEL
+
+        return OVERALL_LABEL
+    if name == "results_to_df":
+        from .results_df import results_to_df
+
+        return results_to_df
+    if name == "hota_results_to_df":
+        from .results_df import hota_results_to_df
+
+        return hota_results_to_df
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
